+++ src/test/groovy/nebula/plugin/rxjavaproject/RxJavaReleasePluginLauncherSpec.groovy
@@ -1,13 +1,27 @@
+/*
+ * Copyright 2014-2015 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * <http://www.apache.org/licenses/LICENSE-2.0>
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package nebula.plugin.rxjavaproject
 
 import com.google.common.io.Files
-import nebula.plugin.publishing.maven.NebulaMavenPublishingPlugin
+import nebula.plugin.publishing.maven.MavenPublishPlugin
 import nebula.test.IntegrationSpec
 import org.ajoberstar.grgit.Grgit
 import org.ajoberstar.grgit.Tag
 import org.ajoberstar.grgit.operation.BranchAddOp
 import org.gradle.api.plugins.JavaPlugin
-import spock.lang.Ignore
 
 class RxJavaReleasePluginLauncherSpec extends IntegrationSpec {
 
@@ -39,7 +53,7 @@ class RxJavaReleasePluginLauncherSpec extends IntegrationSpec {
         buildFile << """
             ext.dryRun = true
             group = 'test'
-            ${applyPlugin(NebulaMavenPublishingPlugin)}
+            ${applyPlugin(MavenPublishPlugin)}
             ${applyPlugin(RxJavaPublishingPlugin)}
             ${applyPlugin(RxJavaReleasePlugin)}
             ${applyPlugin(JavaPlugin)}
@@ -68,7 +82,6 @@ class RxJavaReleasePluginLauncherSpec extends IntegrationSpec {
         !results.wasExecuted(':artifactoryUpload')
         !results.wasUpToDate(':bintrayUpload')
 
-        //Grgit originGit = Grgit.open(origin)
         def tags = originGit.tag.list()
         Tag tag001 = tags.find { Tag tag -> tag.name == 'v0.1.0' }
         tag001
@@ -90,7 +103,6 @@ class RxJavaReleasePluginLauncherSpec extends IntegrationSpec {
     }
 
     def 'perform candidate'() {
-        useToolingApi = false
         when:
         def taskResult = runTasksSuccessfully('tasks', '--all')